Improve zooming in demo.
There is now an arrow pointing where we are zooming in.
The arrow is animated.
The cursor arrow temporarily disappears to make room for this arrow.

And display a number to say how zoomed in we are.

And change the cursor for this animation.
The East-West arrow makes a lot more sense,
sense the user is only controlling left to right motion,
Not up and down.

Tweak the shape of the arrows.