Add a brief pause.
If the animation restarts on its own, then add a pause.
Update the display for the new location and the zoom of 1,
as always.
Then pause for 1/2 second.
Then start zooming.

That 1/2 second pause does not exist when the user clicks.
Or when we start the animation the first time.

I removed a TODO item.
I reorganized some code.