Bug fix:  Arrow was in wrong place.
Regarding the demo where you can move your mouse over the graph
and the big arrow will follow you.
And the arrow disappears when you leave the chart.

In certain cases the arrow would appear,
but the position would be wrong.
As soon as you move the mouse, it would be corrected.

/* HERE */ marks what I was looking at when I figured out this bug.