Create a decent sine wave animation.
It still has glitches to fix.
But is generally looks good.
Lots of comments and TODOs describing this.

I completely changed the interfaces
and implementation for the sine wave.
Originally I wasn't sure how this function would be used.
It's mostly used to draw on charts.
So a period of 2*pi and an amplitude of 1 makes sense.

And add a lot of comments.

Start tearing apart the LC circuit pictures.
That is very rough.

First build in a while.
So all changes will be visible
at https://tradeideasphilip.github.io/divide-by-zero/