Fix the arrow in the first picture
There was a bug in the Pointer class.
Normally the tip of the arrow is exactly at the origin.
But that broke whenever you added a border to the arrow.
It's fixed in this commit.

Small tweaks to the English in the document.

First build in a while.
I've been working on the first animation for a several commits.
It is finished now.