Improve spring animation.
I removed parts of the spring that were in the back.
It's now more clear which lines are in the front,
and which lines are in the back.

Previously the illusion seemed to switch
between one set of lines being in front and then the other.
It was distracting.
And I designed the spring to be seen one way.
The other way was recognizable, but slightly distorted.