Add a missing loop to the spring.
A small part of a loop was missing from the very top.
In some parts of the animation it's completely off the top.

This was an old problem.
The changes in the previous commit made this problem more obvious.

It looks better now!

Reorganize some code.
I was repeating stuff in a loop for no good reason.
It's more efficient now.